pyproject.toml
[tool.poetry]
name = "pymarker"
version = "0.0.0"
description = "A python package to generate AR markers and patterns based on input images"
authors = ["Pablo Silva <pablodiegoss@hotmail.com>"]
readme = "README.md"
homepage = "https://github.com/pablodiegoss/pymarker"
repository = "https://github.com/pablodiegoss/pymarker"
keywords = ["AR", "pattern", "markers", "jandig", "hiro"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent"
]
packages = [
    { include = "pymarker" }
]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.3.0"
click = "^8.2.1"

[tool.poetry.scripts]
pymarker = "pymarker.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
ruff = "^0.12.3"

[tool.pytest.ini_options]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["tests"]

[tool.ruff]
exclude = [
  ".git",
  "migrations",
  ".venv",
]
line-length = 88
# Assume Python 3.12
target-version = "py312"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"